WIP: doclinkchecker
